Moved worlds and characters to own app
