+++ pokerole_irl/user_app/admin.py
@@ -1,14 +1,6 @@
 from django.contrib import admin
 
-from .models import Profile, Setting, Character
+from .models import Profile
 
 admin.site.register(Profile)
-
-@admin.register(Setting)
-class SettingAdmin(admin.ModelAdmin):
-    list_display = ('name', 'owner')
-
-@admin.register(Character)
-class CharacterAdmin(admin.ModelAdmin):
-    list_display = ('first_name', 'last_name', 'owner')
