+++ pokerole_irl/user_app/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.contrib.auth.forms import UserCreationForm, UsernameField
 from django.contrib.auth.models import User
-from .models import Profile, Setting
+from .models import Profile
 
 # Form that creates a user
