+++ pokerole_irl/worlds_app/admin.py
@@ -1,0 +1,11 @@
+from django.contrib import admin
+from .models import Setting, Character
+
+# Register your models here.
+@admin.register(Setting)
+class SettingAdmin(admin.ModelAdmin):
+    list_display = ('name', 'owner')
+
+@admin.register(Character)
+class CharacterAdmin(admin.ModelAdmin):
+    list_display = ('first_name', 'last_name', 'owner')
