+++ pokerole_irl/worlds_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class WorldsAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'worlds_app'
